+++ CHANGELOG.rst
@@ -7,7 +7,8 @@ Features
 * Allow passing ssl context for Twisted (PYTHON-1161)
 * Cloud Twisted support (PYTHON-1163)
 * Add additional_write_policy and read_repair to system schema parsing (PYTHON-1048)
-* Remove *read_repair_chance table options (PYTHON-1140)
+* Remove *read_repair_chance* table options (PYTHON-1140)
+* [GRAPH] Ability to execute Fluent Graph queries asynchronously (PYTHON-1129)
 
 Bug Fixes
 ---------